+++ lib/core/models/wind_entity.dart
@@ -10,7 +10,7 @@ class WindEntity {
   factory WindEntity.fromJson(Map<String, dynamic> json) {
     return WindEntity(
       speed: json['speed'],
-      degress: json['deg'],
+      degress: json['deg'].toDouble(),
     );
   }
 }
